Move unsupported ops calculate to new module and fix empty cases
